Now the cache can be controlled in the video mixer and the video converter. The Video mixer now support blitting draw. Now It's possible to copy parts of an AkVideoPacket. Filters ported: ChangeHSL, Charify, Cinema, ColorFilter, ColorReplace, ColorTap, Contrast, Convolve, DelayGrab, Denoise, Dice, Distort. Added Rotate plugin.

+++ libAvKys/Plugins/ColorFilter/src/colorfilterelement.h
@@ -33,7 +33,7 @@ class ColorFilterElement: public AkElement
                WRITE setColor
                RESET resetColor
                NOTIFY colorChanged)
-    Q_PROPERTY(qreal radius
+    Q_PROPERTY(int radius
                READ radius
                WRITE setRadius
                RESET resetRadius
@@ -54,7 +54,7 @@ class ColorFilterElement: public AkElement
         ~ColorFilterElement();
 
         Q_INVOKABLE QRgb color() const;
-        Q_INVOKABLE qreal radius() const;
+        Q_INVOKABLE int radius() const;
         Q_INVOKABLE bool soft() const;
         Q_INVOKABLE bool disable() const;
 
@@ -69,13 +69,13 @@ class ColorFilterElement: public AkElement
 
     signals:
         void colorChanged(QRgb color);
-        void radiusChanged(qreal radius);
+        void radiusChanged(int radius);
         void softChanged(bool soft);
         void disableChanged(bool disable);
 
     public slots:
         void setColor(QRgb color);
-        void setRadius(qreal radius);
+        void setRadius(int radius);
         void setSoft(bool soft);
         void setDisable(bool disable);
         void resetColor();
